Set up array to object conversion on the ESI type load.
