+++ src/main/java/com/autonomouslogic/everef/refdata/esi/EsiStoreLoader.java
@@ -1,7 +1,6 @@
 package com.autonomouslogic.everef.refdata.esi;
 
 import com.autonomouslogic.everef.refdata.SimpleStoreLoader;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.functions.BiFunction;
@@ -17,7 +16,7 @@ import lombok.extern.log4j.Log4j2;
 public class EsiStoreLoader extends SimpleStoreLoader {
 	@Setter
 	@NonNull
-	private BiFunction<JsonNode, String, ObjectNode> esiTransformer;
+	private BiFunction<ObjectNode, String, ObjectNode> esiTransformer;
 
 	@Setter
 	@NonNull
