+++ src/main/java/com/autonomouslogic/everef/refdata/sde/SdeTypeTransformer.java
@@ -1,8 +1,6 @@
 package com.autonomouslogic.everef.refdata.sde;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.autonomouslogic.everef.refdata.TransformUtil;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.reactivex.rxjava3.functions.Function;
 import javax.inject.Inject;
@@ -13,7 +11,7 @@ import lombok.extern.log4j.Log4j2;
 @Singleton
 public class SdeTypeTransformer implements Function<ObjectNode, ObjectNode> {
 	@Inject
-	protected ObjectMapper objectMapper;
+	protected TransformUtil transformUtil;
 
 	@Inject
 	protected SdeTypeTransformer() {}
@@ -22,32 +20,15 @@ public class SdeTypeTransformer implements Function<ObjectNode, ObjectNode> {
 	public ObjectNode apply(ObjectNode json) throws Throwable {
 		if (json.has("traits")) {
 			var traits = (ObjectNode) json.get("traits");
-			handleArrayToObjectConversion(traits, "misc_bonuses", "importance");
-			handleArrayToObjectConversion(traits, "role_bonuses", "importance");
+			transformUtil.arrayToObject(traits, "misc_bonuses", "importance");
+			transformUtil.arrayToObject(traits, "role_bonuses", "importance");
 			if (traits.has("types")) {
 				var types = (ObjectNode) traits.get("types");
 				types.fields().forEachRemaining(pair -> {
-					handleArrayToObjectConversion(types, pair.getKey(), "importance");
+					transformUtil.arrayToObject(types, pair.getKey(), "importance");
 				});
 			}
 		}
 		return json;
-	}
-
-	private void handleArrayToObjectConversion(ObjectNode root, String field, String keyField) {
-		if (!root.has(field)) {
-			return;
-		}
-		if (!root.get(field).isArray()) {
-			throw new RuntimeException(String.format(
-					"%s is not an array: %s", field, root.get(field).getNodeType()));
-		}
-		var array = (ArrayNode) root.get(field);
-		var obj = objectMapper.createObjectNode();
-		for (JsonNode entry : array) {
-			var key = entry.get(keyField).asText();
-			obj.set(key, entry);
-		}
-		root.set(field, obj);
 	}
 }
