+++ src/test/resources/com/autonomouslogic/everef/refdata/esi/EsiLoaderTest/type-645.json
@@ -2,399 +2,496 @@
   "capacity": 750,
   "description": "The Dominix is one of the old warhorses dating back to the Gallente-Caldari War. While no longer regarded as the king of the hill, it is by no means obsolete. Its formidable hulk and powerful drone arsenal means that anyone not in the largest and latest battleships will regret ever locking horns with it.",
   "dogma_attributes":
-  [
+  {
+    "3":
     {
       "attribute_id": 3,
       "value": 0
     },
+    "4":
     {
       "attribute_id": 4,
       "value": 100250000
     },
+    "9":
     {
       "attribute_id": 9,
       "value": 9350
     },
+    "1547":
     {
       "attribute_id": 1547,
       "value": 3
     },
+    "12":
     {
       "attribute_id": 12,
       "value": 7
     },
+    "13":
     {
       "attribute_id": 13,
       "value": 5
     },
+    "14":
     {
       "attribute_id": 14,
       "value": 6
     },
+    "15":
     {
       "attribute_id": 15,
       "value": 0
     },
+    "18":
     {
       "attribute_id": 18,
       "value": 0
     },
+    "19":
     {
       "attribute_id": 19,
       "value": 1
     },
+    "21":
     {
       "attribute_id": 21,
       "value": 0
     },
+    "37":
     {
       "attribute_id": 37,
       "value": 109
     },
+    "38":
     {
       "attribute_id": 38,
       "value": 750
     },
+    "552":
     {
       "attribute_id": 552,
       "value": 465
     },
+    "48":
     {
       "attribute_id": 48,
       "value": 600
     },
+    "561":
     {
       "attribute_id": 561,
       "value": 10
     },
+    "564":
     {
       "attribute_id": 564,
       "value": 117
     },
+    "55":
     {
       "attribute_id": 55,
       "value": 1100000
     },
+    "2113":
     {
       "attribute_id": 2113,
       "value": 1
     },
+    "11":
     {
       "attribute_id": 11,
       "value": 10000
     },
+    "70":
     {
       "attribute_id": 70,
       "value": 0.1254
     },
+    "524":
     {
       "attribute_id": 524,
       "value": 0.75
     },
+    "76":
     {
       "attribute_id": 76,
       "value": 84000
     },
+    "525":
     {
       "attribute_id": 525,
       "value": 1
     },
+    "600":
     {
       "attribute_id": 600,
       "value": 3
     },
+    "101":
     {
       "attribute_id": 101,
       "value": 0
     },
+    "102":
     {
       "attribute_id": 102,
       "value": 6
     },
+    "1132":
     {
       "attribute_id": 1132,
       "value": 400
     },
+    "109":
     {
       "attribute_id": 109,
       "value": 0.67
     },
+    "110":
     {
       "attribute_id": 110,
       "value": 0.67
     },
+    "111":
     {
       "attribute_id": 111,
       "value": 0.67
     },
+    "1137":
     {
       "attribute_id": 1137,
       "value": 3
     },
+    "1555":
     {
       "attribute_id": 1555,
       "value": 500
     },
+    "633":
     {
       "attribute_id": 633,
       "value": 0
     },
+    "124":
     {
       "attribute_id": 124,
       "value": 16777215
     },
+    "129":
     {
       "attribute_id": 129,
       "value": 1000
     },
+    "1154":
     {
       "attribute_id": 1154,
       "value": 3
     },
+    "136":
     {
       "attribute_id": 136,
       "value": 1
     },
+    "661":
     {
       "attribute_id": 661,
       "value": 1000
     },
+    "662":
     {
       "attribute_id": 662,
       "value": 0.5
     },
+    "153":
     {
       "attribute_id": 153,
       "value": 1.38E-7
     },
+    "1178":
     {
       "attribute_id": 1178,
       "value": 100
     },
+    "1179":
     {
       "attribute_id": 1179,
       "value": 0.01
     },
+    "161":
     {
       "attribute_id": 161,
       "value": 454500
     },
+    "162":
     {
       "attribute_id": 162,
       "value": 250
     },
+    "113":
     {
       "attribute_id": 113,
       "value": 0.67
     },
+    "1196":
     {
       "attribute_id": 1196,
       "value": 0.01
     },
+    "1198":
     {
       "attribute_id": 1198,
       "value": 0.01
     },
+    "1199":
     {
       "attribute_id": 1199,
       "value": 100
     },
+    "1200":
     {
       "attribute_id": 1200,
       "value": 100
     },
+    "182":
     {
       "attribute_id": 182,
       "value": 3336
     },
+    "188":
     {
       "attribute_id": 188,
       "value": 0
     },
+    "192":
     {
       "attribute_id": 192,
       "value": 7
     },
+    "3266":
     {
       "attribute_id": 3266,
       "value": 50
     },
+    "3267":
     {
       "attribute_id": 3267,
       "value": 100
     },
+    "3268":
     {
       "attribute_id": 3268,
       "value": 0.05
     },
+    "1224":
     {
       "attribute_id": 1224,
       "value": 0.5
     },
+    "208":
     {
       "attribute_id": 208,
       "value": 0
     },
+    "209":
     {
       "attribute_id": 209,
       "value": 0
     },
+    "210":
     {
       "attribute_id": 210,
       "value": 22
     },
+    "211":
     {
       "attribute_id": 211,
       "value": 0
     },
+    "217":
     {
       "attribute_id": 217,
       "value": 395
     },
+    "1768":
     {
       "attribute_id": 1768,
       "value": 11331
     },
+    "1259":
     {
       "attribute_id": 1259,
       "value": 0.76
     },
+    "1261":
     {
       "attribute_id": 1261,
       "value": 0.71
     },
+    "1262":
     {
       "attribute_id": 1262,
       "value": 0.79
     },
+    "246":
     {
       "attribute_id": 246,
       "value": 395
     },
+    "1271":
     {
       "attribute_id": 1271,
       "value": 125
     },
+    "1281":
     {
       "attribute_id": 1281,
       "value": 1
     },
+    "263":
     {
       "attribute_id": 263,
       "value": 7920
     },
+    "264":
     {
       "attribute_id": 264,
       "value": 0
     },
+    "265":
     {
       "attribute_id": 265,
       "value": 8800
     },
+    "267":
     {
       "attribute_id": 267,
       "value": 0.5
     },
+    "268":
     {
       "attribute_id": 268,
       "value": 0.9
     },
+    "269":
     {
       "attribute_id": 269,
       "value": 0.65
     },
+    "270":
     {
       "attribute_id": 270,
       "value": 0.65
     },
+    "271":
     {
       "attribute_id": 271,
       "value": 1
     },
+    "272":
     {
       "attribute_id": 272,
       "value": 0.5
     },
+    "273":
     {
       "attribute_id": 273,
       "value": 0.6
     },
+    "274":
     {
       "attribute_id": 274,
       "value": 0.8
     },
+    "277":
     {
       "attribute_id": 277,
       "value": 1
     },
+    "283":
     {
       "attribute_id": 283,
       "value": 375
     },
+    "49":
     {
       "attribute_id": 49,
       "value": 0
     },
+    "2115":
     {
       "attribute_id": 2115,
       "value": 1
     },
+    "422":
     {
       "attribute_id": 422,
       "value": 1
     },
+    "3020":
     {
       "attribute_id": 3020,
       "value": 1
     },
+    "79":
     {
       "attribute_id": 79,
       "value": 8250
     },
+    "479":
     {
       "attribute_id": 479,
       "value": 2500000
     },
+    "482":
     {
       "attribute_id": 482,
       "value": 6000
     },
+    "484":
     {
       "attribute_id": 484,
       "value": 0.75
     },
+    "500":
     {
       "attribute_id": 500,
       "value": 7.5
     },
+    "2045":
     {
       "attribute_id": 2045,
       "value": 1
     }
-  ],
+  },
   "dogma_effects":
-  [
+  {
+    "2186":
     {
       "effect_id": 2186,
       "is_default": false
     },
+    "2187":
     {
       "effect_id": 2187,
       "is_default": false
     },
+    "5427":
     {
       "effect_id": 5427,
       "is_default": false
     },
+    "5428":
     {
       "effect_id": 5428,
       "is_default": false
     },
+    "8377":
     {
       "effect_id": 8377,
       "is_default": false
     }
-  ],
+  },
   "graphic_id": 318,
   "group_id": 27,
   "market_group_id": 81,
